villagefund_backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
python-decouple>=3.8
cloudinary>=1.36

# Production
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9
